fix(firestore): allow reference to be used to query in `Filters`
