+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/field_path_type.dart
