+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/method_channel_query.dart
@@ -224,7 +224,7 @@ class MethodChannelQuery extends QueryPlatform {
   }
 
   @override
-  QueryPlatform whereFilter(Filter filter) {
+  QueryPlatform whereFilter(FilterPlatformInterface filter) {
     return _copyWithParameters(<String, dynamic>{
       'filters': filter.toJson(),
     });
