+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/utils/firestore_message_codec.dart
@@ -8,7 +8,7 @@
 import 'dart:typed_data';
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:cloud_firestore_platform_interface/src/internal/field_path_type.dart';
+import 'package:cloud_firestore_platform_interface/src/field_path_type.dart';
 import 'package:cloud_firestore_platform_interface/src/method_channel/method_channel_field_value.dart';
 import 'package:firebase_core/firebase_core.dart';
 // TODO(Lyokone): remove once we bump Flutter SDK min version to 3.3
