+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/platform_interface/platform_interface_query.dart
@@ -229,7 +229,7 @@ abstract class QueryPlatform extends PlatformInterface {
     throw UnimplementedError('where() is not implemented');
   }
 
-  QueryPlatform whereFilter(Filter filter) {
+  QueryPlatform whereFilter(FilterPlatformInterface filter) {
     throw UnimplementedError('whereFilter() is not implemented');
   }
