+++ packages/cloud_firestore/cloud_firestore_platform_interface/test/field_path_test.dart
@@ -4,7 +4,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:cloud_firestore_platform_interface/src/internal/field_path_type.dart';
+import 'package:cloud_firestore_platform_interface/src/field_path_type.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
