+++ packages/cloud_firestore/cloud_firestore_web/lib/src/query_web.dart
@@ -249,7 +249,7 @@ class QueryWeb extends QueryPlatform {
   }
 
   @override
-  QueryPlatform whereFilter(Filter filter) {
+  QueryPlatform whereFilter(FilterPlatformInterface filter) {
     return _copyWithParameters(<String, dynamic>{
       'filters': filter.toJson(),
     });
